refatorado projeto para formato DDD
