+++ src/main/java/br/com/zupacademy/ggwadera/casadocodigo/autor/AutorRepository.java
@@ -1,6 +1,5 @@
-package br.com.zupacademy.ggwadera.casadocodigo.repository;
+package br.com.zupacademy.ggwadera.casadocodigo.autor;
 
-import br.com.zupacademy.ggwadera.casadocodigo.model.Autor;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface AutorRepository extends JpaRepository<Autor, Long> {
