+++ src/main/java/br/com/zupacademy/ggwadera/casadocodigo/autor/EmailUnicoAutorValidator.java
@@ -1,29 +1,27 @@
-package br.com.zupacademy.ggwadera.casadocodigo.controller.validator;
+package br.com.zupacademy.ggwadera.casadocodigo.autor;
 
-import br.com.zupacademy.ggwadera.casadocodigo.controller.form.AutorForm;
-import br.com.zupacademy.ggwadera.casadocodigo.repository.AutorRepository;
 import org.springframework.stereotype.Component;
 import org.springframework.validation.Errors;
 import org.springframework.validation.Validator;
 
 @Component
-public class EmailUnicoValidator implements Validator {
+public class EmailUnicoAutorValidator implements Validator {
 
     private final AutorRepository repository;
 
-    public EmailUnicoValidator(AutorRepository repository) {
+    public EmailUnicoAutorValidator(AutorRepository repository) {
         this.repository = repository;
     }
 
     @Override
     public boolean supports(Class<?> aClass) {
-        return AutorForm.class.isAssignableFrom(aClass);
+        return NovoAutorDTO.class.isAssignableFrom(aClass);
     }
 
     @Override
     public void validate(Object o, Errors errors) {
         if (errors.hasErrors()) return;
-        final AutorForm form = (AutorForm) o;
+        final NovoAutorDTO form = (NovoAutorDTO) o;
         if (repository.existsByEmail(form.getEmail())) {
             errors.rejectValue("email", null, "O email informado já existe " + form.getEmail());
         }
