+++ src/main/java/br/com/zupacademy/ggwadera/casadocodigo/categoria/CategoriaRepository.java
@@ -1,6 +1,5 @@
-package br.com.zupacademy.ggwadera.casadocodigo.repository;
+package br.com.zupacademy.ggwadera.casadocodigo.categoria;
 
-import br.com.zupacademy.ggwadera.casadocodigo.model.Categoria;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface CategoriaRepository extends JpaRepository<Categoria, Long> {
